виправляє структуру контролерів
