+++ controllers/contacts/index.js
@@ -1,0 +1,16 @@
+const { getAll } = require("./getAll");
+const { getById } = require("./getById");
+const {add} = require("./add");
+const { deleteById } = require("./deleteById");
+const { updateById } = require("./updateById");
+const { updateStatusContact } = require("./updateStatusContact");
+
+module.exports = {
+    getAll,
+    getById,
+    add,
+    deleteById,
+    updateById,
+    updateStatusContact,
+
+}
